chore: split script up into github-action and luarocks-tag-release
