--- entrypoint.sh
@@ -1,5 +1,0 @@
-#!/usr/bin/env bash
-
-nix build ".#luarocks-tag-release"
-
-./result/bin/luarocks-tag-release "$@"
